#47 Reintroduce usecases
- Added usecases for each feature and transferred all providers to it

- Moved prettyPrintJson into logger.dart
