+++ lib/presentation/debug/debug_transactions/providers/debug_transactions_provider.dart
@@ -1,31 +1,8 @@
-import 'package:flutter/foundation.dart';
 import 'package:fpdart/fpdart.dart';
-import 'package:pecunia/core/errors/accounts_errors/accounts_errors.dart';
-import 'package:pecunia/features/accounts/domain/accounts_repo.dart';
 import 'package:pecunia/features/accounts/domain/entities/account.dart';
-import 'package:pecunia/features/auth/domain/auth_repo.dart';
-import 'package:pecunia/features/transactions/domain/entities/transaction.dart';
-import 'package:pecunia/features/transactions/domain/transactions_repo.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'debug_transactions_provider.g.dart';
-
-@riverpod
-class DeleteTransaction extends _$DeleteTransaction {
-  @override
-  Future<Option<Unit>> build() async {
-    return const Option.none();
-  }
-
-  Future<void> deleteTransaction(Transaction txn) async {
-    state = const AsyncValue.loading();
-
-    (await ref.watch(transactionsRepoProvider).deleteTransaction(txn).run()).fold(
-      (l) => state = AsyncError(l, l.stackTrace),
-      (r) => state = AsyncData(Option.of(r)),
-    );
-  }
-}
 
 @riverpod
 class ChosenAccount extends _$ChosenAccount {
@@ -37,59 +14,5 @@ class ChosenAccount extends _$ChosenAccount {
   void updateChosenAccount(List<Account> accountsList, String chosenAccountId) {
     final chosenAccount = accountsList.firstWhere((account) => account.id == chosenAccountId);
     state = Option.of(chosenAccount);
-  }
-}
-
-@riverpod
-class GetAllAccounts extends _$GetAllAccounts {
-  @override
-  Future<List<Account>> build() async {
-    return _getAllAccounts();
-  }
-
-  Future<List<Account>> _getAllAccounts() async {
-    state = const AsyncValue.loading();
-
-    debugPrint('retrieving logged in user...');
-    final failureOrPecuniaUser = await ref.read(authRepoProvider).getLoggedInUser().run();
-
-    final pecuniaUser = failureOrPecuniaUser.fold(
-      (failure) {
-        state = AsyncValue.error(failure, failure.stackTrace);
-      },
-      (pecuniaUser) => pecuniaUser,
-    );
-
-    if (pecuniaUser != null) {
-      debugPrint('user exists, retrieving accounts...');
-      return (await ref.read(accountsRepoProvider).getAccounts().run()).fold(
-        (l) => Future<List<Account>>.error(l, l.stackTrace),
-        (r) => r,
-      );
-    } else {
-      return Future<List<Account>>.error(
-        AccountsFailure.unknown(
-          stackTrace: StackTrace.current,
-          message: 'This should never happen, please contact support.',
-          accountsAction: AccountsAction.unknown,
-        ),
-        StackTrace.current,
-      );
-    }
-  }
-}
-
-@riverpod
-class GetAllTransactions extends _$GetAllTransactions {
-  @override
-  Future<List<Transaction>> build() async {
-    return _getAllTransactions();
-  }
-
-  Future<List<Transaction>> _getAllTransactions() async {
-    return (await ref.read(transactionsRepoProvider).getAllTransactions().run()).fold(
-      (l) => Future<List<Transaction>>.error(l, l.stackTrace),
-      (r) => r,
-    );
   }
 }
